+++ tests/testthat/test-use_start_time.R
@@ -3,8 +3,9 @@ context("use_start_time")
 youtube <- rickroll_youtube()
 youtube_start <- use_start_time(youtube, "25s")
 
-get_query <- function(elem){
-  src <- htmltools::tagGetAttribute(elem, "src")
+get_query <- function(embed) {
+  iframe <- get_iframe(embed)
+  src <- htmltools::tagGetAttribute(iframe, "src")
   url <- httr::parse_url(src)
 
   url$query
@@ -17,8 +18,9 @@ test_that("youtube works", {
 vimeo <- rickroll_vimeo()
 vimeo_start <- use_start_time(vimeo, "1m3s")
 
-get_fragment <- function(elem){
-  src <- htmltools::tagGetAttribute(elem, "src")
+get_fragment <- function(elem) {
+  iframe <- get_iframe(elem)
+  src <- htmltools::tagGetAttribute(iframe, "src")
   url <- httr::parse_url(src)
 
   url$fragment
@@ -26,5 +28,12 @@ get_fragment <- function(elem){
 
 test_that("vimeo works", {
   expect_identical(get_fragment(vimeo_start), "t=63")
+})
+
+test_that("box throws a warning", {
+
+  emb_box <- embed_box("foo")
+
+  expect_warning(use_start_time(emb_box, "25s"), "^Start time")
 })